feat: modularise session CRU(D)
As `loginService` depends on other services that are used within the app, we
often run into circular dependencies. Split the needed methods out into a new
service.

Closes #15.
